Starting Solver class
added: the forder datasets in the Git folder, it is small enough
(<500ko) to be shared
added: solver class
added: class to contain demands and preprocessed data
edit: roster class: only contains the schedule of one nurse now. The
solver will compute and update the costs and the satisfaction of the
constraints.