bidirectional: I'm giving up
It is impossible to understand. When I deactivate improved domination only for consecutive rest, the algorithm exits without error. But now, if I play with the scenario of n030w4 to manually deactivate the constraint on consecutive rest (setting lb=0 and ub=10 for every contract), I get errors again (even if improved domination is deactivated for consecutive rests)...