Bidirectional is working
I finally found the bug!! It was very tricky, and even after finding it, it took me quite some time to understand why it was a bug. I just did the easy fix, I still need to do it better as explained in the big TODO comment of the total weekend bac expander. Sadly, it does not mean that the whole bidirectional thing had a point, since it is still significantly slower.